backend/requirements.txt
# Core Flask and Web
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1

# OpenAI / AI
openai>=1.55.3

# File Processing - NEW for personalized mode
python-docx==1.1.0       # For DOCX text extraction
PyMuPDF>=1.24.0          # For PDF to image conversion (NO poppler dependency needed!)
Pillow>=10.0.0           # Image processing

# Utilities
python-dateutil==2.8.2
gunicorn==21.2.0
